+++ src/Repository/Checkout.php
@@ -16,8 +16,16 @@ use Innmind\Immutable\{
 
 final class Checkout
 {
-    public function __construct(private Binary $binary)
+    private function __construct(private Binary $binary)
     {
+    }
+
+    /**
+     * @internal
+     */
+    public static function of(Binary $binary): self
+    {
+        return new self($binary);
     }
 
     /**